requirements.txt
aiogram==2.5
requests>=2.31.0
python-dotenv
huggingface_hub
